Close #1002 - xsl transformation of the XML when using SMTPSend step
Javascriptable xslFilepath and contentType properties has been added.

If no source selected, the current outputDocument is used, even if no xsl is defined (act as pretty print).

The xsl is apply to the source results (concatened if more than one) or the outputDocument if no source.

The contentType property can be used to force the content type of the mail.

 * without xsl : text/plain
 * with xsl : text/html
 
 The documentation must be reviewed in #3036