Close #3188 - Order parallel step response
Element are now ordered.

Execute from attached project the sequence !WhatTimeParallel and the order must be 8/2/4.

Execute the sequence !IteratorParallel with the test case X3 and X10, the order must be the same as the first part of the XML.

Try in an older CEMS version to compare.

A case doesn't work with the Iterator + Parallel case. Only the first output true step is ordered (set ouput false on the Serial of IteratorParallel). We cannot solve this issue with the current sequencer structure.