closed #4143 - Fixed complexType name generation (use tagname instead of name)
The trouble here was not really about namespace, but about elements with '''same''' name and from '''different''' namespaces.

I've choosen not to 'merge' types from different namespaces because it doesn't mean anything !
Thus, I've prefered to consider the namespace prefix while generating the element's complexType name.
{{{<xsd:complexType name="root_h-tableType">}}} and {{{<xsd:complexType name="root_f-tableType">}}}

But be aware that the generated schema won't be validated :
{{{
cos-element-consistent : Multiple elements with name 'table', with different types, appear in the model group
}}}
meaning it is forbidden under a given element to have children with same name and different types ! [[BR]]

'''Improvement solution (#4623) :'''
 * the XSDExtractor should be rewritten to use !XmlSchema api
 * it should create a collection of schemas, and
  * or, use referenced schema(s) in XML [[BR]]
  see https://msdn.microsoft.com/en-us/library/ms757863(v=vs.85).aspx
  * or, create a schema for each namespace declared in XML