rf #791 Global components created for global uses (Card, AutoGrid..)
Changes on the config page are now being saved using the save button.
Advances have been made in the responsiveness of the console for mobile
screen sizes. The left sidebar is now responsive and transformed into a
burger menu in the mobile version." advancements made on connections and
Projects pages. Skeleton loading added on graphs